SIG-517 Moved email tasks calls to email_integrations app.
