+++ api/app/signals/apps/signals/signal_receivers.py
@@ -1,0 +1,9 @@
+from django.dispatch import receiver
+
+from signals.apps.signals import tasks
+from signals.apps.signals.models import create_initial
+
+
+@receiver(create_initial, dispatch_uid='signals_create_initial')
+def create_initial_handler(sender, signal_obj, **kwargs):
+    tasks.push_to_sigmax.delay(pk=signal_obj.id)
