+++ api/app/signals/apps/signals/templates/email/sigmax/pdf_template.html
